pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "r1_rag"
version = "1.0.0"
description = "Reasoning-First RAG with Global Planning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rag", "llm", "multi-hop-qa", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.4.0",
    "transformers>=4.40.0",
    "sentence-transformers>=2.5.0",
    "networkx>=3.2.0",
    "python-Levenshtein>=0.25.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
train = [
    "vllm>=0.6.3",
    "ray>=2.10.0",
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "flash-attn>=2.5.0",
]
retriever = [
    "faiss-gpu>=1.8.0",
    "pyserini>=0.22.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]
annotation = [
    "openai>=1.12.0",
]
all = ["r1_rag[train,retriever,annotation]"]

[project.urls]
Homepage = "https://github.com/your-username/R1-RAG"
Documentation = "https://github.com/your-username/R1-RAG#readme"
Repository = "https://github.com/your-username/R1-RAG.git"
Issues = "https://github.com/your-username/R1-RAG/issues"

[tool.setuptools.packages.find]
include = ["r1_rag*"]

